+++ src/main/java/com/dailyon/paymentservice/config/RedisConfig.java
@@ -1,6 +1,6 @@
 package com.dailyon.paymentservice.config;
 
-import com.dailyon.paymentservice.domain.payment.paymanger.kakaopay.response.KakaopayReadyResponseVO;
+import com.dailyon.paymentservice.domain.payment.dto.KakaopayReadyDTO;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -33,8 +33,8 @@ public class RedisConfig {
   }
 
   @Bean
-  public RedisTemplate<String, KakaopayReadyResponseVO> redisTemplate() {
-    RedisTemplate<String, KakaopayReadyResponseVO> redisTemplate = new RedisTemplate<>();
+  public RedisTemplate<String, KakaopayReadyDTO> redisTemplate() {
+    RedisTemplate<String, KakaopayReadyDTO> redisTemplate = new RedisTemplate<>();
     redisTemplate.setKeySerializer(new StringRedisSerializer());
     redisTemplate.setValueSerializer(new GenericJackson2JsonRedisSerializer());
     redisTemplate.setConnectionFactory(redisConnectionFactory());
