+++ src/main/java/com/dailyon/paymentservice/domain/client/KakaopayFeignClient.java
@@ -1,7 +1,7 @@
 package com.dailyon.paymentservice.domain.client;
 
 import com.dailyon.paymentservice.config.KakaopayFeignConfig;
-import com.dailyon.paymentservice.domain.payment.paymanger.kakaopay.response.KakaopayReadyResponseVO;
+import com.dailyon.paymentservice.domain.payment.dto.KakaopayReadyDTO;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.util.MultiValueMap;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -14,6 +14,6 @@ import org.springframework.web.bind.annotation.RequestHeader;
     configuration = KakaopayFeignConfig.class)
 public interface KakaopayFeignClient {
   @PostMapping("/v1/payment/ready")
-  KakaopayReadyResponseVO ready(
+  KakaopayReadyDTO ready(
       @RequestHeader(value = "Authorization") String authorization, @RequestBody MultiValueMap map);
 }
