+++ src/main/java/com/dailyon/paymentservice/domain/payment/dto/MemberPointUpdateDTO.java
@@ -1,15 +1,15 @@
-package com.dailyon.paymentservice.domain.client.request;
+package com.dailyon.paymentservice.domain.payment.dto;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor
-public class MemberPointUpdateRequest {
+public class MemberPointUpdateDTO {
   private static final String source = "CHARGE";
   private Long amount;
 
-  public MemberPointUpdateRequest(Long amount) {
+  public MemberPointUpdateDTO(Long amount) {
     this.amount = amount;
   }
 }
