+++ src/main/java/com/dailyon/paymentservice/domain/payment/facades/PaymentFacade.java
@@ -1,16 +1,11 @@
 package com.dailyon.paymentservice.domain.payment.facades;
 
-import com.dailyon.paymentservice.domain.client.MemberFeignClient;
-import com.dailyon.paymentservice.domain.client.request.MemberPointUpdateRequest;
 import com.dailyon.paymentservice.domain.payment.api.request.PaymentReadyRequest;
-import com.dailyon.paymentservice.domain.payment.entity.Payment;
 import com.dailyon.paymentservice.domain.payment.paymanger.KakaoPayManager;
-import com.dailyon.paymentservice.domain.payment.paymanger.kakaopay.response.KakaopayReadyResponseVO;
-import com.dailyon.paymentservice.domain.payment.service.PaymentService;
+import com.dailyon.paymentservice.domain.payment.dto.KakaopayReadyDTO;
 import com.dailyon.paymentservice.domain.payment.utils.OrderNoGenerator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
@@ -20,7 +15,7 @@ public class PaymentFacade {
   public String pointPaymentReady(
       Long memberId, PaymentReadyRequest.PointPaymentReadyRequest request) {
     String orderId = OrderNoGenerator.generate(memberId);
-    KakaopayReadyResponseVO readyResponse = kakaoPayManager.ready(memberId, orderId, request);
+    KakaopayReadyDTO readyResponse = kakaoPayManager.ready(memberId, orderId, request);
     return readyResponse.getNextRedirectAppUrl();
   }
 }
