+++ src/main/java/com/dailyon/paymentservice/domain/payment/repository/RedisRepository.java
@@ -1,6 +1,6 @@
 package com.dailyon.paymentservice.domain.payment.repository;
 
-import com.dailyon.paymentservice.domain.payment.paymanger.kakaopay.response.KakaopayReadyResponseVO;
+import com.dailyon.paymentservice.domain.payment.dto.KakaopayReadyDTO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Repository;
@@ -11,9 +11,9 @@ import java.time.Duration;
 @RequiredArgsConstructor
 public class RedisRepository {
 
-  private final RedisTemplate<String, KakaopayReadyResponseVO> redisTemplate;
+  private final RedisTemplate<String, KakaopayReadyDTO> redisTemplate;
 
-  public void saveReadyInfo(String key, KakaopayReadyResponseVO data) {
+  public void saveReadyInfo(String key, KakaopayReadyDTO data) {
     redisTemplate.opsForValue().set(key, data, Duration.ofMinutes(5));
   }
 }
